requirements.txt
opencv-python
imread-from-url
onnxruntime
cap-from-youtube
pymongo
python-dotenv
cloudinary
moviepy
vidgear